chore(suite): rename, move, delete various wallet components
