+++ packages/suite/src/components/suite/modals/TransactionDetail/components/ChangeFee/components/RbfFees.tsx
@@ -1,4 +1,4 @@
-import { Fees } from 'src/components/wallet/Fees';
+import { Fees } from 'src/components/wallet/Fees/Fees';
 import { useRbfContext } from 'src/hooks/wallet/useRbfForm';
 
 // wrapper for shareable Fees component
