+++ packages/suite/src/components/wallet/index.tsx
@@ -1,9 +1,7 @@
-import Title from './Title';
 import { WalletLayout } from './WalletLayout/WalletLayout';
 import { WalletLayoutHeader } from './WalletLayout/WalletLayoutHeader';
 import { WalletLayoutNavigation } from './WalletLayout/WalletLayoutNavigation';
 import { WalletLayoutNavLink } from './WalletLayout/WalletLayoutNavLink';
-import OnOffSwitcher from './OnOffSwitcher';
 import { InputError } from './InputError';
 import { AccountExceptionLayout } from './AccountExceptionLayout';
 import { DiscoveryProgress } from './DiscoveryProgress';
@@ -15,14 +13,12 @@ import type { WithSelectedAccountLoadedProps } from './hocs';
 import { CoinjoinAccountDiscoveryProgress } from './CoinjoinAccountDiscoveryProgress/CoinjoinAccountDiscoveryProgress';
 
 export {
-    Title,
     WalletLayout,
     WalletLayoutHeader,
     WalletLayoutNavigation,
     WalletLayoutNavLink,
     DiscoveryProgress,
     withSelectedAccountLoaded,
-    OnOffSwitcher,
     InputError,
     AccountExceptionLayout,
     UtxoAnonymity,
