+++ packages/suite/src/views/wallet/coinmarket/exchange/components/ExchangeForm/Fees/index.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Fees } from 'src/components/wallet/Fees';
+import { Fees } from 'src/components/wallet/Fees/Fees';
 import { useCoinmarketExchangeFormContext } from 'src/hooks/wallet/useCoinmarketExchangeForm';
 
 const StyledCard = styled.div`
