+++ packages/suite/src/views/wallet/coinmarket/sell/components/SellForm/Fees/index.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Fees } from 'src/components/wallet/Fees';
+import { Fees } from 'src/components/wallet/Fees/Fees';
 import { useCoinmarketSellFormContext } from 'src/hooks/wallet/useCoinmarketSellForm';
 
 const StyledCard = styled.div`
