+++ packages/suite/src/views/wallet/send/components/SendFees.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import { variables } from '@trezor/components';
 
 import { Card } from 'src/components/suite';
-import { Fees } from 'src/components/wallet/Fees';
+import { Fees } from 'src/components/wallet/Fees/Fees';
 import { useSendFormContext } from 'src/hooks/wallet';
 
 const StyledCard = styled(Card)`
